requirements.txt
flask
flask-cors
sounddevice
"numpy<2"
httpx
piper-tts
soundfile
python-multipart
pyttsx3


# DB VECTORIELLE
chromadb
pydantic==1.10.13

# DB RECHERCHE DE CONNAISSANCES
faiss-cpu
sentence-transformers
